// src/generated/projectRegistry.ts
/* This file is generated by scripts/generate-project-registry.mjs. */
import type { InfographicTemplate } from "../layoutCatalog";
import type { TranscriptPage } from "../video/types";
import template_process_optimization from "../projects/process-optimization/template.json";
import tokens_process_optimization from "../projects/process-optimization/tokens.json";
import video_process_optimization from "../projects/process-optimization/video.mp4";
import template_test_project from "../projects/test-project/template.json";

export type ProjectRegistryEntry = {
  captions: boolean;
  hasTokens: boolean;
  hasVideo: boolean;
  name: string;
  template: InfographicTemplate;
  transcriptPages: TranscriptPage[];
  videoBased: boolean;
  videoSrc?: string;
};

export const projectRegistry = [
  {
    hasTokens: true,
    hasVideo: true,
    name: "process-optimization",
    template: template_process_optimization as InfographicTemplate,
    transcriptPages: tokens_process_optimization as TranscriptPage[],
    videoSrc: video_process_optimization,
    videoBased: true,
    captions: true,
  },
  {
    hasTokens: false,
    hasVideo: false,
    name: "test-project",
    template: template_test_project as InfographicTemplate,
    transcriptPages: [],
    videoSrc: undefined,
    videoBased: false,
    captions: false,
  }
] satisfies ProjectRegistryEntry[];